+++ volume-cartographer/core/include/vc/fiber_tracer/FiberPaths.hpp
@@ -176,6 +176,9 @@ struct FiberletPathReport {
     size_t retainedSearchNodes = 0;
     size_t interpolationCornerInsertions = 0;
     size_t interpolatedScoringPoints = 0;
+    size_t scoringPageCount = 0;
+    size_t scoringPageSlots = 0;
+    size_t scoringPageDirectoryProbes = 0;
     size_t dpNodeIndexEntries = 0;
     size_t dpNodeIndexSlots = 0;
     size_t dpTransitionLookups = 0;